SweetAlert2, validate fields

+++ crud-firebase-bootstrap/src/app/pages/employees/details/details.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { NavigationExtras, Router } from '@angular/router';
 import { EmployeeI } from 'src/app/shared/models/employee.interface';
 import { EmployeesService } from 'src/app/shared/services/employees.service';
+import Swal from "sweetalert2";
 
 @Component({
   selector: 'app-details',
@@ -42,10 +43,29 @@ export class DetailsComponent implements OnInit {
   }
 
   async deleteEmployee(id: string) {
-    await this.employeesService.deleteEmployee(id).then(() => {
-      console.log('Deleted employee...');
-      this.onGoToBack();
-    }).catch(error => console.log('Error delete...', error));
+    await Swal.fire({
+      title: '¿Está seguro?',
+      text: "¡No podrás revertir esto!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: '¡Sí, eliminar!',
+      cancelButtonText: 'Cancelar',
+    }).then(async result => {
+      if (result.isConfirmed) {
+        await this.employeesService.deleteEmployee(id).then(() => {
+          Swal.fire(
+            '¡Eliminado!',
+            'El registro ha sido eliminado.',
+            'success',
+          );
+          console.log('Deleted employee...');
+          this.onGoToBack();
+        }).catch(error => console.log('Error delete...', error));
+      }
+    });
+
   }
